move MouseState and related classes to mouse_state.py
