+++ system/ui/widgets/button.py
@@ -3,7 +3,8 @@ from enum import IntEnum
 
 import pyray as rl
 
-from openpilot.system.ui.lib.application import FontWeight, MousePos
+from openpilot.system.ui.lib.application import FontWeight
+from openpilot.system.ui.lib.mouse_state import MousePos
 from openpilot.system.ui.widgets import Widget
 from openpilot.system.ui.widgets.label import Label
